move l-bfgs-b tests into test_minimize_constr.py
