+++ tests/torchmin/test_minimize_constr.py
@@ -197,3 +197,102 @@ def test_frankwolfe_tracenorm():
     )
     est_X = res.x
     torch.testing.assert_close(est_X, 0.2 * eye, rtol=1e-2, atol=1e-2)
+
+
+def test_lbfgsb_simple_quadratic():
+    """Test L-BFGS-B on a simple bounded quadratic problem.
+
+    Minimize: f(x) = (x1 - 2)^2 + (x2 - 1)^2
+    Subject to: 0 <= x1 <= 1.5, 0 <= x2 <= 2
+
+    The unconstrained minimum is at (2, 1), but x1 is constrained,
+    so the optimal solution should be at (1.5, 1).
+    """
+
+    def fun(x):
+        return (x[0] - 2)**2 + (x[1] - 1)**2
+
+    x0 = torch.tensor([0.5, 0.5])
+    lb = torch.tensor([0.0, 0.0])
+    ub = torch.tensor([1.5, 2.0])
+
+    result = minimize_constr(
+        fun,
+        x0,
+        method='l-bfgs-b',
+        bounds=(lb, ub),
+        options=dict(gtol=1e-6, ftol=1e-9),
+    )
+
+    # Check if close to expected solution
+    expected_x = torch.tensor([1.5, 1.0])
+    expected_f = 0.25
+
+    assert torch.allclose(result.x, expected_x, atol=1e-4), \
+        f"Solution {result.x} not close to expected {expected_x}"
+    assert abs(result.fun - expected_f) < 1e-4, \
+        f"Function value {result.fun} not close to expected {expected_f}"
+
+
+def test_lbfgsb_rosenbrock():
+    """Test L-BFGS-B on Rosenbrock function with bounds.
+
+    Minimize: f(x,y) = (1-x)^2 + 100(y-x^2)^2
+    Subject to: -2 <= x <= 2, -2 <= y <= 2
+
+    The unconstrained minimum is at (1, 1).
+    """
+
+    x0 = torch.tensor([-1.0, 1.5])
+    lb = torch.tensor([-2.0, -2.0])
+    ub = torch.tensor([2.0, 2.0])
+
+    result = minimize_constr(
+        rosen,
+        x0,
+        method='l-bfgs-b',
+        bounds=(lb, ub),
+        options=dict(gtol=1e-6, ftol=1e-9, max_iter=100),
+    )
+
+    # Check if close to expected solution
+    expected_x = torch.tensor([1.0, 1.0])
+
+    assert torch.allclose(result.x, expected_x, atol=1e-3), \
+        f"Solution {result.x} not close to expected {expected_x}"
+    assert result.fun < 1e-6, \
+        f"Function value {result.fun} not close to 0"
+
+
+def test_lbfgsb_active_constraints():
+    """Test L-BFGS-B with multiple active constraints.
+
+    Minimize: f(x) = sum(x_i^2)
+    Subject to: x_i >= 1 for all i
+
+    The solution should be all ones (on the boundary).
+    """
+
+    def fun(x):
+        return (x**2).sum()
+
+    n = 5
+    x0 = torch.ones(n) * 2.0
+    lb = torch.ones(n)
+    ub = torch.ones(n) * 10.0
+
+    result = minimize_constr(
+        fun,
+        x0,
+        method='l-bfgs-b',
+        bounds=(lb, ub),
+        options=dict(gtol=1e-6, ftol=1e-9),
+    )
+
+    expected_x = torch.ones(n)
+    expected_f = float(n)
+
+    assert torch.allclose(result.x, expected_x, atol=1e-4), \
+        f"Solution {result.x} not close to expected {expected_x}"
+    assert abs(result.fun - expected_f) < 1e-4, \
+        f"Function value {result.fun} not close to expected {expected_f}"
